requirements.txt
torch
torchvision
timm
matplotlib
numpy
pandas
tqdm
Pillow
kagglehub